Fixed cast warning on GNUstep
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@7996 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/ETOutlineLayout.m
@@ -373,7 +373,7 @@ expanded and collapsed by getting automatically a related outline arrow. */
 
 		[outlineView setDropItem: dropItem dropChildIndex: dropIndex];
 
-		ETLog(@"Retarget drop to %ld in %@", (NSUInteger)dropIndex, dropItem);
+		ETLog(@"Retarget drop to %ld in %@", (long)dropIndex, dropItem);
 	}
 
 	return NSDragOperationEvery;
